+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This package follows standard semvar, `<major>.<minor>.<build>`. No breaking cha
 * Fix issues with allOf/oneOf with partially completed data.
 * Sanitize paths coming from the spec that contain invalid characters.
 * Improve display of array types and number/string formats in parameters and models
+* Fix fetch request options to not require unnecessary extra level in setting properties.
 
 ## 0.8 ##
 * Removing the `RESET` button as it is confusing for users to see. It only repopulated the defaults, and to do that, you can easily switch tabs and come back.